lagt till superklass
